refactor: Clean up package imports and rename layer directories to plural for consistency
