+++ src/main/java/com/vann/controllers/CustomerController.java
@@ -1,28 +1,17 @@
-package com.vann.controller;
+package com.vann.controllers;
 
 import java.net.URI;
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.*;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import com.vann.exceptions.FieldConflictException;
-import com.vann.exceptions.RecordNotFoundException;
-import com.vann.model.Customer;
-import com.vann.service.CustomerService;
+import com.vann.exceptions.*;
+import com.vann.models.Customer;
+import com.vann.services.CustomerService;
 import com.vann.utils.LogHandler;
