+++ src/main/java/com/vann/controllers/InvoiceController.java
@@ -1,20 +1,14 @@
-package com.vann.controller;
+package com.vann.controllers;
 
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.*;
+import org.springframework.web.bind.annotation.*;
 
 import com.vann.exceptions.RecordNotFoundException;
-import com.vann.model.Invoice;
-import com.vann.service.InvoiceService;
+import com.vann.models.Invoice;
+import com.vann.services.InvoiceService;
 import com.vann.utils.LogHandler;
 
 @RestController
