+++ src/main/java/com/vann/models/Cart.java
@@ -1,18 +1,10 @@
-package com.vann.model;
+package com.vann.models;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.UUID;
+import java.util.*;
 
 import com.vann.utils.LogHandler;
 
-import jakarta.persistence.CollectionTable;
-import jakarta.persistence.Column;
-import jakarta.persistence.ElementCollection;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.MapKeyColumn;
+import jakarta.persistence.*;
 
 @Entity
 public class Cart {
