+++ src/main/java/com/vann/models/Invoice.java
@@ -1,16 +1,11 @@
-package com.vann.model;
+package com.vann.models;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 import java.util.regex.Pattern;
 
 import com.vann.utils.LogHandler;
 
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
+import jakarta.persistence.*;
 
 @Entity
 public class Invoice {
