+++ src/main/java/com/vann/models/enums/Size.java
@@ -1,4 +1,4 @@
-package com.vann.model.enums;
+package com.vann.models.enums;
 
 public enum Size {
