+++ src/main/java/com/vann/repositories/CategoryRepo.java
@@ -1,14 +1,12 @@
 package com.vann.repositories;
 
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.vann.model.Category;
-import com.vann.model.enums.CategoryType;
+import com.vann.models.Category;
+import com.vann.models.enums.CategoryType;
 
 @Repository
 public interface CategoryRepo extends JpaRepository<Category, UUID> {
