+++ src/main/java/com/vann/repositories/CustomerRepo.java
@@ -1,12 +1,11 @@
 package com.vann.repositories;
 
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.vann.model.Customer;
+import com.vann.models.Customer;
 
 @Repository
 public interface CustomerRepo extends JpaRepository<Customer, UUID> {
