+++ src/main/java/com/vann/repositories/InvoiceRepo.java
@@ -1,12 +1,11 @@
 package com.vann.repositories;
 
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.vann.model.Invoice;
+import com.vann.models.Invoice;
 
 @Repository
 public interface InvoiceRepo extends JpaRepository<Invoice, UUID> {
