+++ src/main/java/com/vann/repositories/ProductRepo.java
@@ -1,12 +1,11 @@
 package com.vann.repositories;
 
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.vann.model.Product;
+import com.vann.models.Product;
 
 @Repository
 public interface ProductRepo extends JpaRepository<Product, UUID> {
