+++ src/main/java/com/vann/services/CategoryService.java
@@ -1,15 +1,13 @@
-package com.vann.service;
+package com.vann.services;
 
-import com.vann.exceptions.RecordNotFoundException;
-import com.vann.exceptions.FieldConflictException;
-import com.vann.model.Category;
-import com.vann.model.enums.CategoryType;
-import com.vann.repositories.CategoryRepo;
+import java.util.*;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
+import com.vann.exceptions.*;
+import com.vann.models.Category;
+import com.vann.models.enums.CategoryType;
+import com.vann.repositories.CategoryRepo;
+
 
 @Service
 public class CategoryService {
