+++ src/test/java/com/vann/models/CartTest.java
@@ -1,13 +1,10 @@
-package com.vann.model;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.UUID;
+package com.vann.models;
 
 import static org.junit.jupiter.api.Assertions.*;
+import org.junit.jupiter.api.*;
+
+import java.util.*;
+
 
 public class CartTest {
