+++ src/test/java/com/vann/models/CategoryTest.java
@@ -1,13 +1,12 @@
-package com.vann.model;
+package com.vann.models;
 
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import com.vann.model.enums.CategoryType;
+import static org.junit.jupiter.api.Assertions.*;
+import org.junit.jupiter.api.*;
 
 import java.util.UUID;
 
-import static org.junit.jupiter.api.Assertions.*;
+import com.vann.models.enums.CategoryType;
+
 
 public class CategoryTest {
