+++ src/test/java/com/vann/models/InvoiceTest.java
@@ -1,14 +1,12 @@
-package com.vann.model;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.UUID;
+package com.vann.models;
 
 import static org.junit.jupiter.api.Assertions.*;
+
+import java.util.*;
+
+import org.junit.jupiter.api.*;
+import org.mockito.Mockito;
+
 
 public class InvoiceTest {
