+++ src/test/java/com/vann/services/CartServiceTest.java
@@ -1,32 +1,18 @@
 package com.vann.services;
 
-import com.vann.exceptions.RecordNotFoundException;
-import com.vann.model.Cart;
-import com.vann.model.Category;
-import com.vann.model.Customer;
-import com.vann.model.Product;
-import com.vann.repositories.CartRepo;
-import com.vann.repositories.CustomerRepo;
-import com.vann.service.CartService;
-import com.vann.service.CustomerService;
-import com.vann.service.ProductService;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.UUID;
-
 import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.*;
+
+import java.util.*;
+
+import org.junit.jupiter.api.*;
+import org.mockito.*;
+
+import com.vann.exceptions.RecordNotFoundException;
+import com.vann.models.*;
+import com.vann.repositories.*;
+
 
 public class CartServiceTest {
