+++ src/test/java/com/vann/services/InvoiceServiceTest.java
@@ -1,32 +1,17 @@
 package com.vann.services;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyInt;
+import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.*;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
 
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
+import org.junit.jupiter.api.*;
+import org.mockito.*;
 
 import com.vann.exceptions.RecordNotFoundException;
-import com.vann.model.Customer;
-import com.vann.model.Invoice;
-import com.vann.model.InvoiceItem;
+import com.vann.models.*;
 import com.vann.repositories.InvoiceRepo;
-import com.vann.service.CustomerService;
-import com.vann.service.InvoiceItemService;
-import com.vann.service.InvoiceService;
 
 
 public class InvoiceServiceTest {
